+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to **DecaLog** is documented in this *changelog*.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and **DecaLog** adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- The PHP listener can now detect upgrading/downgrading, extensions activation/deactivation.
+
 ## [1.1.1] - 2019-08-29
 ### Changed
 - Better 'Page not found' detection and report.