+++ CHANGELOG.md
@@ -5,8 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 ### Added
-- The PHP listener can now detect upgrading/downgrading, extensions activation/deactivation.
-- The database listener can now detect upgrading/downgrading.
+- The WordPress listener can now detect version upgrading/downgrading.
+- The PHP listener can now detect version upgrading/downgrading.
+- The PHP listener can now detect extensions activation/deactivation.
+- The database listener can now detect version upgrading/downgrading.
 
 ## [1.1.1] - 2019-08-29
 ### Changed