+++ CHANGELOG.md
@@ -18,6 +18,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Some events have new levels and/or messages to comply with rules described in `DEVELOPER.md`.
 ### Fixed
 - WordPress formatter may (wrongly) emit warnings when source ip is unknown.
+### Security
+- In a multisite, a non-admin user was able to read the name of a logger (not its content).
 
 ## [1.1.0] - 2019-08-28
 ### Added