+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - The PHP listener can now detect version upgrading/downgrading.
 - The PHP listener can now detect extensions activation/deactivation.
 - The database listener can now detect version upgrading/downgrading.
+_ The plugin now embeds its own inline help (help tab in all screens).
 ### Security
 - In a multisite, a non-admin user was able to read the name of a logger (not its content).