+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to **DecaLog** is documented in this *changelog*.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and **DecaLog** adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.2.0] - 2019-08-31
 ### Added
 - The WordPress listener can now detect version upgrading/downgrading.
 - The PHP listener can now detect version upgrading/downgrading.